Update i915 PXP and MEI Driver to DII client.
Tracked-On: OAM-120419

+++ drivers/gpu/drm/i915/gt/iov/intel_iov_ggtt.c
@@ -173,7 +173,7 @@ static bool is_pte_replicable(struct intel_iov *iov, gen8_pte_t pte)
 void intel_iov_ggtt_vf_update_pte(struct intel_iov *iov, u32 offset, gen8_pte_t pte)
 {
 	struct intel_iov_vf_ggtt_ptes *buffer = &iov->vf.ptes_buffer;
-	u16 max_copies = FIELD_MAX(VF2PF_MMIO_UPDATE_GGTT_REQUEST_MSG_1_NUM_COPIES);
+	u16 max_copies = (u8)FIELD_MAX(VF2PF_MMIO_UPDATE_GGTT_REQUEST_MSG_1_NUM_COPIES);
 	u16 max_ptes = MMIO_UPDATE_GGTT_MAX_PTES;
 	u32 pte_offset = (offset >> PAGE_SHIFT) - (iov->vf.config.ggtt_base >> PAGE_SHIFT);
